// src/dotnet/Blazor3D/Blazor3D.TypeScript/src/Generated/blazor3d-configuration.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

import { SceneOptions_B3D } from "./scene-options";
import { ViewerOptions_B3D } from "./viewer-options";
import { CameraOptions_B3D } from "./camera-options";
import { OrbitControlOptions_B3D } from "./orbit-control-options";

export class Blazor3DConfiguration_B3D {
    scene: SceneOptions_B3D;
    sceneOrtho: SceneOptions_B3D;
    viewerOptions: ViewerOptions_B3D;
    cameraOptions: CameraOptions_B3D;
    orbitControlOptions: OrbitControlOptions_B3D;
}