// src/dotnet/Blazor3D/Blazor3D.TypeScript/src/Generated/custom-geometry.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

import { BufferGeometry_B3D } from "./buffer-geometry";

/**
 * <para>Class for rectangular cuboid with a given 'width', 'height', and 'depth'.</para>
 * <para>This class inherits from @see {@link Blazor3D.Core.BufferGeometry}</para>
 * <para>Wrapper for three.js <a target="_blank" href="https://threejs.org/docs/index.html#api/en/geometries/Buffer2Geometry">Buffer2Geometry</a></para>
 * 
 * <inheritdoc>@see {@link Blazor3D.Core.BufferGeometry}</inheritdoc>
 */
export class CustomGeometry_B3D extends BufferGeometry_B3D {
    vertexBuffer: string;
    indexBuffers: string[] = [];
}